Flytter tilbake fil med responsobjekter som bare brukes av nav veileder og konverterer mockdata motsatt vei for å unngå å flytte typer til feil modul pga mocker
