+++ apps/innbyggers-flate/src/mocks/MockHandler.ts
@@ -16,12 +16,12 @@ import {
   getLedetekst,
   getUtvidetInnhold,
   harBakgrunnsinfo,
-  lagHistorikkFellesOppstart
+  lagHistorikkFellesOppstart,
+  lagOpplaringKategoriseringMockRespons
 } from 'deltaker-flate-common'
 import { HttpResponse } from 'msw'
 import { v4 as uuidv4 } from 'uuid'
 import { DeltakerResponse } from '../api/data/deltaker.ts'
-import { lagOpplaringKategoriseringMockRespons } from 'deltaker-flate-common'
 
 const bakgrunnsinformasjon =
   'Ønsker å bli kontaktet via sms\nKan ikke på onsdager'
