+++ packages/deltaker-flate-common/index.ts
@@ -31,7 +31,6 @@ export * from './model/endre-deltaker'
 export * from './model/forslag'
 export * from './model/utils'
 export * from './model/kodeverk'
-export * from './model/kodeverkResponse'
 export * from './model/prisinformasjon'
 export * from './utils/displayText'
 export * from './utils/constants'
